# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Commune API",
    "version": "1.0.0",
    "description": "The Commune API lets AI agents and developers send and receive email, manage inboxes and domains, and process structured data from email conversations. All endpoints require a Bearer API key.\n\nBase URL: `https://api.commune.email`\n\nAuthentication: Pass your API key as `Authorization: Bearer comm_...`",
    "contact": {
      "name": "Commune Support",
      "url": "https://commune.email",
      "email": "support@commune.email"
    },
    "license": {
      "name": "Business Source License 1.1",
      "url": "https://mariadb.com/bsl11/"
    }
  },
  "servers": [
    {
      "url": "https://api.commune.email",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "comm_...",
        "description": "API key issued from the Commune dashboard. Prefix `comm_`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
      "Participant": {
        "type": "object",
        "required": ["role", "identity"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["sender", "to", "cc", "bcc", "mentioned", "participant"],
            "description": "Participant's role in the message"
          },
          "identity": {
            "type": "string",
            "description": "Email address of the participant"
          }
        }
      },
      "MessageMetadata": {
        "type": "object",
        "properties": {
          "created_at": { "type": "string", "format": "date-time" },
          "subject": { "type": "string" },
          "in_reply_to": { "type": "string", "nullable": true },
          "references": { "type": "array", "items": { "type": "string" } },
          "inbox_id": { "type": "string", "nullable": true },
          "inbox_address": { "type": "string", "nullable": true },
          "domain_id": { "type": "string", "nullable": true },
          "message_id": { "type": "string", "nullable": true },
          "delivery_status": {
            "type": "string",
            "enum": ["sent", "delivered", "bounced", "failed", "complained", "suppressed", "blocked", "credit_limit"],
            "nullable": true
          },
          "delivery_data": {
            "type": "object",
            "nullable": true,
            "properties": {
              "sent_at": { "type": "string", "format": "date-time" },
              "delivered_at": { "type": "string", "format": "date-time" },
              "bounced_at": { "type": "string", "format": "date-time" },
              "failed_at": { "type": "string", "format": "date-time" },
              "bounce_reason": { "type": "string" },
              "bounce_type": { "type": "string", "enum": ["hard", "soft"] },
              "failure_reason": { "type": "string" }
            }
          },
          "has_attachments": { "type": "boolean" },
          "attachment_count": { "type": "integer" },
          "attachment_ids": { "type": "array", "items": { "type": "string" } },
          "spam_checked": { "type": "boolean" },
          "spam_score": { "type": "number", "minimum": 0, "maximum": 1 },
          "spam_action": { "type": "string", "enum": ["reject", "flag", "accept"] },
          "spam_flagged": { "type": "boolean" },
          "prompt_injection_detected": { "type": "boolean" },
          "prompt_injection_risk": {
            "type": "string",
            "enum": ["none", "low", "medium", "high", "critical"]
          },
          "extracted_data": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "description": "Structured data extracted via inbox extraction schema (if configured)"
          }
        }
      },
      "Message": {
        "type": "object",
        "required": ["message_id", "thread_id", "channel", "direction", "participants", "content", "created_at", "metadata"],
        "properties": {
          "message_id": { "type": "string", "description": "Unique message identifier" },
          "thread_id": { "type": "string", "description": "Thread this message belongs to" },
          "channel": { "type": "string", "enum": ["email", "sms"], "description": "Communication channel" },
          "direction": { "type": "string", "enum": ["inbound", "outbound"] },
          "participants": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Participant" }
          },
          "content": { "type": "string", "description": "Plain text content" },
          "content_html": { "type": "string", "nullable": true, "description": "HTML content" },
          "attachments": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of attachment IDs"
          },
          "created_at": { "type": "string", "format": "date-time" },
          "metadata": { "$ref": "#/components/schemas/MessageMetadata" }
        }
      },
...
      "SendEmailRequest": {
        "type": "object",
        "required": ["to", "subject"],
        "properties": {
          "to": {
            "oneOf": [
              { "type": "string", "format": "email" },
              { "type": "array", "items": { "type": "string", "format": "email" }, "minItems": 1 }
            ],
            "description": "Recipient(s)"
          },
          "subject": { "type": "string", "maxLength": 500 },
          "html": { "type": "string", "description": "HTML body. At least one of html or text is required." },
          "text": { "type": "string", "description": "Plain text body. At least one of html or text is required." },
          "from": { "type": "string", "format": "email", "description": "Sender address (overrides inbox default)" },
          "cc": {
            "oneOf": [
              { "type": "string", "format": "email" },
              { "type": "array", "items": { "type": "string", "format": "email" } }
            ]
          },
          "bcc": {
            "oneOf": [
              { "type": "string", "format": "email" },
              { "type": "array", "items": { "type": "string", "format": "email" } }
            ]
          },
          "reply_to": { "type": "string", "format": "email" },
          "thread_id": { "type": "string", "description": "Thread to reply into" },
          "inbox_id": { "type": "string", "description": "Inbox to send from (recommended)" },
          "domain_id": { "type": "string", "description": "Domain to send from (inferred from inbox_id if provided)" },
          "attachments": {
            "type": "array",
            "items": {
              "oneOf": [
                { "type": "string", "description": "Attachment ID from POST /v1/attachments/upload" },
                {
                  "type": "object",
                  "required": ["filename", "content"],
                  "properties": {
                    "filename": { "type": "string" },
                    "content": { "type": "string", "description": "Base64-encoded content" },
                    "content_type": { "type": "string" }
                  }
                }
              ]
            }
          },
          "headers": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Custom email headers"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "error": "Rate limit exceeded" }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  },
  "paths": {
    "/v1/messages/send": {
      "post": {
        "tags": ["Messages"],
        "operationId": "sendEmail",
        "summary": "Send an email",
        "description": "Send an outbound email message. Supports threading, CC/BCC, HTML/text bodies, and file attachments. Emails are queued via BullMQ and delivered through your domain's SMTP configuration.\n\nRequires `messages:write` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Unique key for idempotent request deduplication. If a request with the same key was already processed within 24 hours, the cached response is returned with header Idempotency-Replayed: true.",
            "schema": { "type": "string", "example": "550e8400-e29b-41d4-a716-446655440000" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendEmailRequest" },
              "example": {
                "to": "alice@example.com",
                "subject": "Hello from Commune",
                "html": "<p>Hello!</p>",
                "inbox_id": "550e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Email accepted and queued for delivery",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "description": "Pre-generated message ID (msg_...)", "example": "msg_abc123def456" },
                        "status": { "type": "string", "enum": ["queued"], "description": "Delivery status" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/v1/messages": {
      "get": {
        "tags": ["Messages"],
        "operationId": "listMessages",
        "summary": "List messages",
        "description": "List messages filtered by inbox, domain, or sender. At least one filter is required.\n\nRequires `messages:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "inbox_id",
            "in": "query",
            "description": "Filter by inbox ID",
            "schema": { "type": "string" }
          },
          {
            "name": "domain_id",
            "in": "query",
            "description": "Filter by domain ID",
            "schema": { "type": "string" }
          },
          {
            "name": "sender",
            "in": "query",
            "description": "Filter by sender email address",
            "schema": { "type": "string", "format": "email" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (1–1000, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Return messages before this ISO date",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Return messages after this ISO date",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Message" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/v1/threads/{threadId}/messages": {
      "get": {
        "tags": ["Threads"],
        "operationId": "getThreadMessages",
        "summary": "Get messages in a thread",
        "description": "Returns all messages in a thread in chronological order.\n\nRequires `threads:read` permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max messages to return (1–1000, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order (oldest first by default)",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages in the thread",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Message" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
...
}